Cleanup polling and reduce duplicate code
